Allow .pro files to copy files
It is possible that a .pro file that uses our .pri would also want to make use of our “hack” to get common .java files into the android build folder.

In order to support this, we now check to see if `first` has previously been declared and extend it, otherwise we behave like the did before.